requirements.txt
# 网络设备连接
netmiko>=4.2.0
paramiko>=3.4.0
napalm>=4.1.0

# 命令输出解析
textfsm>=1.1.3
ntc-templates>=4.0.1

# 数据处理和存储
pandas>=2.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0

# 配置和日志
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0

# 并发处理
concurrent-log-handler>=0.9.24

# 进度显示
tqdm>=4.66.0
